incorporate exchanges and do a refactor to make some other changes easier
